Added message to .gitignore
As the message says. There is a plethora of resources that have contradictory information on whether it's okay to include google-services.json in a public GitHub, so I decided to add it to this file for now. Please reach out to me if you need the file for the app to function.