Password Checking Features Implementation
After some web surfing, I found a solution to check for strong passwords (using code from here: https://medium.com/@kamranbekirovyz/building-a-customizable-password-strength-checker-in-flutter-ac219650305a). I also rearranged the checking logic to properly halt the user when they enter a poor password. The registration page is now able to check that the password is decently strong and that the entered password matches the re-entered one.