Firestore functions setup
I added a new function that adds the newly registered user information to the Firestore database for this project. I'm slowly working through how to attach the tasks a user sets to a given user (i.e., storing information with Firestore), but it's taking longer than anticipated to detangle it.