Completed User Registration
I completed the user registration feature using Firestore (a customizable database in Firebase). The code correctly submits the two fields (email and username) to the "users" database. I also started the process of overhauling the task page to allow for storing tasks in the "tasks" database (which is attached to a user)

I also downgraded the Firestore package because the newer version caused problems with compilation.