Initial Database Connections
I have some database (Firestore) capabilities to add a task, but there is a massive issue where the task is not being associated with a user (it's being added like IT is a user).

I'm still trying to work out how to handle the user's tasks. One option is to keep the local storage (Hive) and update the tasks within those functions. The other option is to remove the Hive functions and replace them with Firestore functions. I need to get the Firestore functions working first before getting a more solid conclusion on which option works best.