Firebase Firestore Connection - Adding Tasks
I finally figured out a way to add tasks as a sub-collection under a user. I'm currently working out the task deletion (it's not working because I have to know the ID of the task to remove it). It's possible the best way to handle task deletion (and task retrieval) is to rewrite the way the task list is updated to include Firestore snapshots (based on the resources I've looked at).