Fine-Tuning Settings Page
I realize that I won't be able to complete (or if I do, not cleanly) the task management in the database, so I switched gears to build out the settings page. Still trying to figure out how to have more items in a list rather than crammed in one box. I had to cut off the coding for the night or else I would have ended up asleep at the keyboard.