Task List Retrieval w/ Firebase
I revamped the functions in "firestore.dart" (which will eventually replace the giant chunk of function in "tasks_page.dart") for a more streamlined look and added the get tasks function. I also started a new task page to implement the new Firebase functions in. The hope is that if this new page works, I can transfer the code into the current task page file (basically cutting out the Hive portion and using only Firebase).