Settings Page & Firebase Integration
I created a basic settings page based on a YouTube tutorial. The functions don't go anywhere at this time, but it at least looks pretty. I haven't been able to figure out how to get the current user's username to appear in place of "USER," so that's just going to be a thing for now.

The bigger achievement is that the app can properly add, update, and delete tasks in Firestore and show the list of tasks a user has!!! I'm still working to re-integrate the checkbox (to change the completion status, which will be reflected with the "completed" field in the database), but I should be working on branch merging in one or two commits here.