Integration Testing
I'm testing strategies to implement the Firebase capabilities of my branch to the current main branch, which still uses Hive. The current roadblock is the way the Task class is implemented. I just don't have the time or experience to create a new Task class with Firebase. I'll open an issue with that and go from there.