Old Login Page Integration Test (+ Page Rearrangement)
I tested how the new navigation would react with the user profile page system (under UserPage) would react with the new UI. Seems fine so far, but I did have to make minor changes to the dependencies. I also changed the arrangement of the pages so the user page is the primary page and the others (register, login, auth) were part of the "auth" folder.