Messing with Settings Functionality
Tried to wrestle with adding settings to the app. Added some localizations in the main file, but haven't found a clear way to implement them in the settings. Also trying to discern how the app reacts to accessibility settings, but the emulator either isn't employing them right or the app is unaffected by them (i.e., I need to test it on a physical phone).