Updating Firestore Task Functions
I'm beginning the process of updating the current functions I made in finalizeLogin_ND to accept Task objects and pass them to the database. I've also been re-evaluating what will be saved in the database that comes from the task. I currently have the task name, task description, completion status, and deadline date directly translated to the database. I will need to think about handling the clock in/out status and find a way to translate the task labels.