Account Information Update; Language Setting Setup
I officially got the general account option to correctly present the user's username and email when the option is selected (although there is an initial delay where you need to click the option once). The spacing is all weird in the dialog box, so that's something to work on.
I also started setting up the language setting dialog box. They're only buttons for now, but the hope is that the static text in the app will be translated (using the assumption that the user will enter task information in different languages).
I have changed the "Notification" section to an "Acknowledgement" (name pending) section, where I think the credits and privacy statement should go.