Code Cleanup
I cleaned up the current Profile Page (i.e., removed unused functions and touched up the text). I still don't know why the dialog boxes are so big, but we're rolling with it for now. I'm going to remove the Firebase-Firestore functions for this branch and save them for a different branch in the next commit.