Added App Icon/Logo; Profile Page Changes
I added an icon for the app, which is a sword (with colors based on the app's) driven through a task list. I was hoping to make this the app's image instead of the default Flutter icon, but I can't figure that out right now.

I also made the following changes to the profile page:
- Removed unused code
- Adjusted function names
- Added a small privacy statement
- Removed dark mode in-app function (for now)