Email Verification Patch
After the 5/2 presentation, I pondered if my understanding on how Firebase's email verification function worked like I said it would. I checked and discovered someone could bypass the email verification page by restarting the page. So, after a few hours of wrestling with Flutter to get the pages to work, I found a solution that prompts the email verification page again if a person restarts the app.

THERE IS A SLIGHT CAVEAT!! If you are in debug mode and you navigate away from and back to the verification page within the app, it stops due to an error. I tried running the app without debug mode and I was able to run it normally, so I'm not 100% sure what was happening. This is at least a warning for anyone testing the login functionality later.